refactor(ptv3): move BackbonePreprocessor to its own file
Relocate the BackbonePreprocessor class from ptv3_trt.{hpp,cpp} into
backbone_preprocessor.{hpp,cpp}. Pure move plus the include and CMake source
entry; PTv3TRT is unchanged.
